apps/api/requirements.txt
fastapi>=0.109.0
Pillow>=10.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
uvicorn[standard]>=0.27.0
